New folder (5)/requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
loguru==0.7.2
flask==3.0.0
flask-socketio==5.3.6
eventlet==0.33.3
